Python/Courses/DRU3/07_Flask/requirements.txt
SQLAlchemy
Flask
Flask_SQLAlchemy
psycopg2
python-dotenv